+++ src/app.c
@@ -1,5 +1,6 @@
 #include "editor/util.h"
 #include "app.h"
+#include "editor/widget.h"
 #include "events.h"
 #include "gfx/keys.h"
 #include "te_dbg.h"
@@ -7,6 +8,7 @@
 #include <bits/time.h>
 #include <stdbool.h>
 #include <stdio.h>
+#include <string.h>
 #include <time.h>
 
 TE_App te_app_init(int w, int h) {
@@ -16,6 +18,12 @@ TE_App te_app_init(int w, int h) {
 
     // TODO: prepare basic widgets
 
+    TE_Label* lbl = te_label_new("Hello, world!");
+    lbl->widget.x = 50;
+    lbl->widget.y = 100;
+
+    TE_VecP_Widget_push(&app.widgets, &lbl->widget);
+
     return app;
 }
 
@@ -23,9 +31,9 @@ void te_app_run(TE_App* app) {
     long tgt_fps = 60;
     long tgt_frame_time = (long)(1000.0 / (double)tgt_fps);
     DBG_PRINT("TGT FPS: %lu, TGT FRAME TIME: %lu\n", tgt_fps, tgt_frame_time);
-
-    bool running = true;
-    while (running) {
+    
+    app->running = true;
+    while (app->running) {
         struct timespec start_time;
         clock_gettime(CLOCK_MONOTONIC, &start_time);
 
@@ -41,6 +49,12 @@ void te_app_run(TE_App* app) {
                     );
                     break;
                 }
+                case TE_KeyRelease: {
+                    if (te.val.key.key == TE_KEY_ESCAPE) {
+                        app->running = false;
+                    }
+                    break;
+                }
                 case TE_InputText: {
                     DBG_PRINT("Text input %s\n",
                         te.val.key.text
@@ -51,9 +65,19 @@ void te_app_run(TE_App* app) {
                     break;
             }
             
+            // TODO: send event to widget if needed
 
             te = gfx_poll();
         }
+
+        for (size_t i = 0; i < app->widgets.len; i++) {
+            TE_Widget* widget = app->widgets.data[i];
+            if (widget->visible) {
+                widget->draw(widget);
+            }
+        }
+
+        gfx_flush(); 
 
         struct timespec end_time;
         clock_gettime(CLOCK_MONOTONIC, &end_time);
@@ -63,6 +87,14 @@ void te_app_run(TE_App* app) {
         long elapsed_ms = (long)(elapsed_time * 1000.0);
         if (elapsed_ms < tgt_frame_time) {
             TE_Sleep(tgt_frame_time - elapsed_ms);
+        }
+    }
+
+    for (size_t i = 0; i < app->widgets.len; i++) {
+        TE_Widget* widget = app->widgets.data[i];
+       
+        if (widget->dtor) {
+            widget->dtor(widget);
         }
     }